rename jobs module to pipelines
