+++ conf/deployment.yml
@@ -32,14 +32,14 @@ environments:
       - name: 'telco-churn-demo-setup'
         <<: *dev-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/demo_setup_job.py'
+          python_file: 'file://telco_churn/pipelines/demo_setup_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.dev.env',
                        '--conf-file', 'file:fuse://conf/job_configs/demo_setup.yml']
       - name: 'telco-churn-feature-table-creation'
         <<: *dev-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/feature_table_creator_job.py'
+          python_file: 'file://telco_churn/pipelines/feature_table_creator_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.dev.env',
                        '--conf-file', 'file:fuse://conf/job_configs/feature_table_creation.yml']
@@ -47,7 +47,7 @@ environments:
         <<:
           - *dev-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_train_job.py'
+          python_file: 'file://telco_churn/pipelines/model_train_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.dev.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_train.yml']
@@ -55,7 +55,7 @@ environments:
         <<:
           - *dev-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_deployment_job.py'
+          python_file: 'file://telco_churn/pipelines/model_deployment_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.dev.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_deployment.yml']
@@ -63,7 +63,7 @@ environments:
         <<:
           - *dev-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_inference_job.py'
+          python_file: 'file://telco_churn/pipelines/model_inference_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.dev.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_inference_batch.yml']
@@ -93,7 +93,7 @@ environments:
       - name: 'telco-churn-demo-setup'
         <<: *prod-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/demo_setup_job.py'
+          python_file: 'file://telco_churn/pipelines/demo_setup_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.prod.env',
                        '--conf-file', 'file:fuse://conf/job_configs/demo_setup.yml' ]
@@ -103,7 +103,7 @@ environments:
             <<:
               - *prod-cluster-config
             spark_python_task:
-              python_file: 'file://telco_churn/jobs/demo_setup_job.py'
+              python_file: 'file://telco_churn/pipelines/demo_setup_job.py'
               parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                            '--env', 'file:fuse://conf/.prod.env',
                            '--conf-file', 'file:fuse://conf/job_configs/demo_setup.yml' ]
@@ -112,7 +112,7 @@ environments:
             depends_on:
               - task_key: 'demo-setup'
             spark_python_task:
-              python_file: 'file://telco_churn/jobs/feature_table_creator_job.py'
+              python_file: 'file://telco_churn/pipelines/feature_table_creator_job.py'
               parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                            '--env', 'file:fuse://conf/.prod.env',
                            '--conf-file', 'file:fuse://conf/job_configs/feature_table_creation.yml']
@@ -122,7 +122,7 @@ environments:
               - task_key: 'demo-setup'
               - task_key: 'feature-table-creation'
             spark_python_task:
-              python_file: 'file://telco_churn/jobs/model_train_job.py'
+              python_file: 'file://telco_churn/pipelines/model_train_job.py'
               parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                            '--env', 'file:fuse://conf/.prod.env',
                            '--conf-file', 'file:fuse://conf/job_configs/model_train.yml']
@@ -130,7 +130,7 @@ environments:
         <<:
           - *prod-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_train_job.py'
+          python_file: 'file://telco_churn/pipelines/model_train_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.prod.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_train.yml']
@@ -138,7 +138,7 @@ environments:
         <<:
           - *prod-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_deployment_job.py'
+          python_file: 'file://telco_churn/pipelines/model_deployment_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.prod.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_deployment.yml']
@@ -146,7 +146,7 @@ environments:
         <<:
           - *prod-cluster-config
         spark_python_task:
-          python_file: 'file://telco_churn/jobs/model_inference_job.py'
+          python_file: 'file://telco_churn/pipelines/model_inference_job.py'
           parameters: ['--base-data-params', 'file:fuse://conf/.base_data_params.env',
                        '--env', 'file:fuse://conf/.prod.env',
                        '--conf-file', 'file:fuse://conf/job_configs/model_inference_batch.yml']
